% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMicrostep.R
\name{addMicrostep}
\alias{addMicrostep}
\title{A function to create the new edgelist resulting from a microstep in \pkg{RSiena}}
\usage{
addMicrostep(dat, newedge)
}
\arguments{
\item{dat}{\code{data.frame} The edgelist that is changing. Must have a "from" column and a "to" column of vertex IDs.}

\item{newedge}{A named vector of the new edge in the network. The nodes must be represented in the same way as in dat, and the names "from" and "to" must exist}
}
\description{
A function to create the new edgelist resulting from a microstep in \pkg{RSiena}
}
