% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorEdges.R
\name{colorEdges}
\alias{colorEdges}
\title{Function to color the edges in a microstep animation}
\usage{
colorEdges(edges_unnested, colorall = "grey20", colorchange = "red",
  sizeall = 0.5, sizechange = 2.5)
}
\arguments{
\item{edges_unnested}{\code{data.frame} Edge data that has been merged with nested node data and unnested}

\item{colorall}{\code{character} Character value of color for all edges when they're not changing}

\item{colorchange}{\code{character} Character value of color for the edges when changing}

\item{sizeall}{\code{numeric} Size of all edges when they're not changing}

\item{sizechange}{\code{numeric} Size of the edge to emphasize the change}
}
\description{
Function to color the edges in a microstep animation
}
