% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorVertices.R
\name{colorVertices}
\alias{colorVertices}
\title{Coloring vertices for animation}
\usage{
colorVertices(vertexdata, pte, colorall = "grey40", colorchange = "red",
  sizeall = 1, sizechange = 5)
}
\arguments{
\item{vertexdata}{\code{data.frame}}

\item{pte}{\code{list} List of microsteps result from \code{\link{pretween_edges}}}

\item{colorall}{\code{character} Character value of color for all nodes when they're not changing}

\item{colorchange}{\code{character} Character value of color for the ego nodes when changing}

\item{sizeall}{\code{numeric} Size of all nodes when they're not changing}

\item{sizechange}{\code{numeric} Size of the ego node to emphasize the change}
}
\description{
Coloring vertices for animation
}
