% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-lineup.R
\name{create_lineup}
\alias{create_lineup}
\title{Create a function to randomly place a data plot among null plots.}
\usage{
create_lineup(sims, dat, sd = NULL)
}
\arguments{
\item{sims}{data frame from the sims_to_df function. Only one wave at a time permitted, so it must be subset. Must have columns from, to, and sim at least.}

\item{dat}{the data you wish to plot among the sims randomly. Must have same column names as the sims data. The sim column must have the value "dat".}

\item{sd}{seed. Use for setting a random seed for reproducibility. Defaults to NULL.}
}
\description{
Create a function to randomly place a data plot among null plots.
}
\examples{
M1sims5df <- sims_to_df(M1sims5)
wave2 <- read.table("data/s50-network2.dat")
wave2$from <- 1:50
tidyr::gather(wave2, key = to, value = tie, V1:V50) \%>\%
dplyr::mutate(to = readr::parse_number(to)) \%>\%
dplyr::filter(tie == 1) -> wave2df
wave2df <- merge(wave2df, data.frame(id = 1:50), by.x = "from", by.y = "id", all = T)
wave2df$tie <- "dat"
names(wave2df)[3] <- "sim"
M1sims5df <- M1sims5df[c(1,2,5)]
test <- create_lineup(M1sims5df, wave2df)

}
