% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crete-net-animate.R
\name{create_net_animate}
\alias{create_net_animate}
\title{Construct plots to be animated.}
\usage{
create_net_animate(dat)
}
\arguments{
\item{dat}{\code{data.frame} The output of a call to \code{\link{tween_microsteps}}}
}
\description{
Construct plots to be animated.
}
