% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-sig-effs.R
\name{find_sig_effs}
\alias{find_sig_effs}
\title{Find significant effects added to the simple RSiena model.}
\usage{
find_sig_effs(dat, N = 1000)
}
\arguments{
\item{N}{Integer. The number of simulations to run in the RSiena fitting structure. Default to 1000.}

\item{RSiena}{data object. A fully constructed RSiena data object, including covariates.}
}
\description{
Find significant effects added to the simple RSiena model.
}
\examples{
source("subset-friendship.R")
testfn <- netvizinf:::find_sig_effs(dat = mydataa)

}
