% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTrueEdges.R
\name{findTrueEdges}
\alias{findTrueEdges}
\title{Take all possible edges and a given edgelist and "remove" the edges that aren't actually there by making the from and to node coordinates identical.}
\usage{
findTrueEdges(alledgesLayout, edgelist)
}
\arguments{
\item{edgelist}{\code{data.frame} Edgelist of network. Must have columns named "from" and "to"}

\item{alledges}{\code{data.frame} Output from the \code{\link{layoutAll()}} function}
}
\description{
Take all possible edges and a given edgelist and "remove" the edges that aren't actually there by making the from and to node coordinates identical.
}
