% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-chain-info.R
\name{get_chain_info}
\alias{get_chain_info}
\title{Function to extract chains from a \code{\link[RSiena]{sienaFit}} object}
\usage{
get_chain_info(ans)
}
\arguments{
\item{ans}{\code{sienaFit} A \code{\link[RSiena]{sienaFit}} object resulting from a \code{\link[RSiena]{siena07}} call.}
}
\description{
Function to extract chains from a \code{\link[RSiena]{sienaFit}} object
}
