% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-changes.R
\name{get_changes}
\alias{get_changes}
\title{A function to get a data frame of vertices that change in the microsteps}
\usage{
get_changes(pte)
}
\arguments{
\item{pte}{\code{list} A list of microstep edgelists. Created from \code{\link{pretween_edges}}}
}
\description{
A function to get a data frame of vertices that change in the microsteps
}
