% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-effects.R
\name{get_effects}
\alias{get_effects}
\title{Create effects creation function for use with purrr}
\usage{
get_effects(dat, eff_basic)
}
\arguments{
\item{dat}{data frame of the effects to include. Columns must be named shortName, type and inter1.}
}
\description{
Create effects creation function for use with purrr
}
