% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMicrostepsDF.R
\name{getMicrostepsDF}
\alias{getMicrostepsDF}
\title{Get a data frame of microstep changes}
\usage{
getMicrostepsDF(microsteps)
}
\arguments{
\item{microsteps}{\code{list} A list of data frames resulting from \code{\link{listMicrosteps}}}
}
\value{
A data frame of microsteps with the following columns:
\itemize{
   \item \code{step} The microstep in which the change occurs
   \item \code{from} The ego node changing a tie
   \item \code{to} The alter node to which the tie from the ego node is changing
   \item \code{edge.weight} Unimportant. Is NA if an edge is added
   \item \code{type} Either "add" or "remove" indicating whether the (from,to) edge is being added or removed in that microstep
}
}
\description{
Get a data frame of microstep changes
}
