% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinData.R
\name{joinData}
\alias{joinData}
\title{Function to join the edgelist and \code{tween}'d layout information. To be used on a \code{tween}'d data frame \code{\link[tidyr]{nest}}ed by microstep and frame.}
\usage{
joinData(currentlayout, currentMS, microsteps)
}
\arguments{
\item{currentlayout}{\code{data.frame} Current layout of the network. Column names should be (at least) x, y, id, addedge, rmvedge}

\item{currentMS}{\code{numeric} Current microstep}

\item{microsteps}{\code{list} Result of \code{\link{pretween_edges}}}
}
\description{
Function to join the edgelist and \code{tween}'d layout information. To be used on a \code{tween}'d data frame \code{\link[tidyr]{nest}}ed by microstep and frame.
}
