% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layoutAll.R
\name{layoutAll}
\alias{layoutAll}
\title{A function to join a layout of a network with all possible edges in that network}
\usage{
layoutAll(layout, alledges)
}
\arguments{
\item{layout}{\code{data.frame} A data frame of a network layout, created from \code{\link{seedLayout}}}

\item{alledges}{\code{data.frame} A data frame with "from" and "to" named columns that contains all possible nodes in a network}
}
\description{
A function to join a layout of a network with all possible edges in that network
}
