% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listMicrosteps.R
\name{listMicrosteps}
\alias{listMicrosteps}
\title{Take an initial network edgelist and a data frame of microsteps and create a list of edgelists of the same length as the number of microsteps + 1.}
\usage{
listMicrosteps(dat, microsteps)
}
\arguments{
\item{dat}{The initial observed network edgelist. Must have a "from" column and a "to" column of vertex IDs.}

\item{microsteps}{A dataframe of output from \link{get_chain_info()}. Should be subset to the period of interest and have columns named "from" and "to".}
}
\description{
Take an initial network edgelist and a data frame of microsteps and create a list of edgelists of the same length as the number of microsteps + 1.
}
