% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net-avg.R
\name{net_avg}
\alias{net_avg}
\title{Average network from a simulation}
\usage{
net_avg(sims, wave, th)
}
\arguments{
\item{sims}{Output from saom_simulate.}

\item{wave}{integer. Which wave of the network would you like to average?}

\item{th}{numeric. Threshold for edge inclusion i.e. how many times must an edge occur to be included in the "average"}
}
\description{
Average network from a simulation
}

