% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretween-edges.R
\name{pretween_edges}
\alias{pretween_edges}
\title{A function that takes a list of microsteps creates some variables for cool visualization}
\usage{
pretween_edges(microsteps)
}
\arguments{
\item{microsteps}{\code{list} A list of microsteps from \code{\link{listMicrosteps}}.}
}
\description{
A function that takes a list of microsteps creates some variables for cool visualization
}
