% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretween-vertices.R
\name{pretween_vertices}
\alias{pretween_vertices}
\title{Create a list of data frames of network microsteps able to be \code{tween}'d.}
\usage{
pretween_vertices(microsteps, layoutparams)
}
\arguments{
\item{microsteps}{\code{list} A list of microsteps of class \code{"data.frame"} of the continuous time Markov chain from one network observation to another.}

\item{layoutParams}{\code{list} A named list of arguments to pass to \code{\link{seedLayout()}}. Needs elements named "n" and "directed" at minimum.}
}
\description{
Create a list of data frames of network microsteps able to be \code{tween}'d.
}
