% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saom-simulate.R
\name{saom_simulate}
\alias{saom_simulate}
\title{Function for simulating from a stochastic actor-oriented model.}
\usage{
saom_simulate(dat, struct, parms, N)
}
\arguments{
\item{dat}{A Siena data object. For more information see ?RSiena::sienaDataCreate}

\item{struct}{A Siena effects structure object. Include all effects you wish to include in the model. See ?RSiena::getEffects for more.}

\item{parms}{Numeric vector of effects. The rate parameters must be listed first, followed by the effects in the objective function.}

\item{N}{Integer. Number of simulations you desire from the model.}
}
\description{
Function for simulating from a stochastic actor-oriented model.
}

