% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedLayout.R
\name{seedLayout}
\alias{seedLayout}
\title{A function to take an edglist and lay it out given a seed}
\usage{
seedLayout(edgelist, n, algorithm = "spring", layout.par = NULL,
  seed = 9384750, from = "from", to = "to")
}
\arguments{
\item{edgelist}{\code{data.frame} Edgelist with "from" and "to" columns. Vertext names should be of the form "V1", "V2",...,"Vn".}

\item{n}{\code{int} Total number of edges in the network}

\item{algorithm}{\code{char} Algorithm from \pkg{sna} to use. See \code{?\link[sna]{gplot.layout}} for options}

\item{layout.par}{\code{list} List of layout parameters.}

\item{seed}{\code{int} Seed to use for layout algorithm}

\item{from}{\code{char} Name of the "from" column}

\item{to}{\code{char} Name of the "to" column}
}
\description{
A function to take an edglist and lay it out given a seed
}
