% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tween_microsteps.R
\name{tween_microsteps}
\alias{tween_microsteps}
\title{A function to create the full set of \code{\link[tweenr]{tween}}'d microstep data appropriate for animation.}
\usage{
tween_microsteps(ptv, pte, microsteps_df)
}
\arguments{
\item{ptv}{\code{list} A list of vertices and their locations at each microstep. The result of \code{\link{pretween_vertices}}}

\item{pte}{\code{list} A list of edges and their locations at each microstep. The result of \code{\link{pretween_edges}}}

\item{microsteps_df}{\code{data.frame}}
}
\value{
A data frame with the following columns:
\itemize{
    \item \code{from} ego node
    \item \code{to} alter node
    \item \code{edge.weight} not important
    \item \code{addedge} is the edge being added
    \item \code{rmvedge} is the edge being removed
    \item \code{microstep} the microstep (0 is the starting network)
    \item \code{from.x} from node x coordinate
    \item \code{from.y} from node y coordinate
    \item \code{ms} animation microstep
    \item \code{size} node size in the frame
    \item \code{color} node color in the frame
    \item \code{.frame} frame id for animating
    \item \code{to.x} to node x coordinate
    \item \code{to.y} to node y coordinate
    \item \code{ecolor} edge color in the frame
    \item \code{esize} edge size in the frame
}
}
\description{
A function to create the full set of \code{\link[tweenr]{tween}}'d microstep data appropriate for animation.
}
