% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tween-removeEdge.R
\name{tween_removeEdge}
\alias{tween_removeEdge}
\title{A function to compute \code{\link[tweenr]{tween}}'d data for an edge removal in a microstep process animation}
\usage{
tween_removeEdge(nodes1, nodes2, edges, stepInfo)
}
\arguments{
\item{nodes1}{\code{data.frame} The node placement for the current microstep}

\item{nodes2}{\code{data.frame} The node placement for the next microstep}

\item{edges}{\code{data.frame} The edge configuration for the current microstep}

\item{stepInfo}{\code{vector} The current step information as a named vector, as a subset of the result of \code{\link{getMicrostepsDF}}}
}
\value{
A data frame with the following columns:
\itemize{
    \item \code{from} ego node
    \item \code{to} alter node
    \item \code{edge.weight} not important
    \item \code{addedge} is the edge being added
    \item \code{rmvedge} is the edge being removed
    \item \code{microstep} the microstep (0 is the starting network)
    \item \code{from.x} from node x coordinate
    \item \code{from.y} from node y coordinate
    \item \code{ms} animation microstep
    \item \code{size} node size in the frame
    \item \code{color} node color in the frame
    \item \code{.frame} frame id for animating
    \item \code{to.x} to node x coordinate
    \item \code{to.y} to node y coordinate
    \item \code{ecolor} edge color in the frame
    \item \code{esize} edge size in the frame
}
}
\description{
A function to compute \code{\link[tweenr]{tween}}'d data for an edge removal in a microstep process animation
}
