% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net-avg.R
\name{net_avg}
\alias{net_avg}
\title{Summarize networks simulated from a SAO model by counting edge occurences}
\usage{
net_avg(sims, wave)
}
\arguments{
\item{sims}{Output from saom_simulate.}

\item{wave}{integer. Which wave of the network would you like to average?}
}
\description{
Summarize networks simulated from a SAO model by counting edge occurences
}
